requirements.txt
pandas~=1.4.1
numpy~=1.22.2
PyYAML~=5.4.1
requests~=2.26.0
Flask~=2.0.2
colorlog~=6.6.0
scikit-image~=0.19.2
Pillow~=9.1.0
torch~=1.11.0
scipy~=1.8.0
prettytable~=3.3.0
easydict~=1.9
torchvision~=0.12.0
timm~=0.5.4
scikit-learn~=1.0.2
matplotlib~=3.5.2
progressbar~=2.5
hexhamming~=2.2.3
opencv-python~=4.5.5.64
boto3~=1.21.10
awscli~=1.22.65